requirements.txt
setuptools
paho.mqtt
minimalmodbus
